+++ tests/Content/CronjobLineTest.php
@@ -146,6 +146,13 @@ class CronjobLineTest extends TestCase
         self::assertEquals(1, $cronjobLine->getDateTimeDefinition()->getMonths()[0]->getValueFrom());
     }
 
+    public function testWeekdayAbbreviations(): void
+    {
+        $line = '* * * * fri www-data command';
+        $cronjobLine = new CronJobLine($line, true);
+        self::assertEquals(5, $cronjobLine->getDateTimeDefinition()->getWeekdays()[0]->getValueFrom());
+    }
+
     private function checkEmptyValues(CronJobLine $cronjobLine): void
     {
         $dateTimeDefinition = $cronjobLine->getDateTimeDefinition();